refactor: ListItem, ListVew 리팩토링
- 둘 다 태그 하나만 포함하므로 atoms로 이동
- ListItem.js: span -> li 변경
- ListView.js: ol, ul 옵션 추가
- theme.js: ol, ul 옵션 관련 상수 추가
